flask/requirements.txt
flask
pickle-mixin
sklearn
numpy
pandas
gunicorn
tensorflow-cpu